Extracted poisson to separate file, now working fully
